Refactoring into JS modules
