+++ public/index.html
@@ -3,6 +3,7 @@
 
 <head>
     <meta charset="utf-8">
+    <meta name="viewport" content="initial-scale=1,width=device-width">
     <title>Look around you: Photos from nearby</title>
     <link rel="stylesheet" href="./style.css">
 </head>
@@ -11,8 +12,21 @@
     <h1>Look around you: photos from nearby</h1>
     <main>
         <article>
-            <h2>You are in <span class="place loader"></span></h2>
+
+
             <section id="screen1" class="active">
+                <form id="placeChooser">
+                    <fieldset>
+                        <legend>Choose a place</legend>
+                        <label><input checked type="radio" name="place" value="here">Use your current location</label>
+                        <label><input type="radio" id="choosePlace" name="place" value="choosePlace">Use a location you
+                            enter</label>
+                        <div class="inactive" id="placeInput"><label>Enter the name of a
+                                place<input type="text" id="userInputPlace"></label></div>
+                        <button type="submit">Submit</button>
+                    </fieldset>
+                </form>
+                <h2>You are in <span class="place loader"></span></h2>
                 <div class="actions">
                     <button id="fetchPhotos" type="button">Find recent photos from <span
                             class="place loader"></span></button>
@@ -40,12 +54,15 @@
                     <button id="startOver" type="button">Start over</button>
                     <div><button class="inactive" id="zip" type="button">Get your ZIP code</button></div>
 
+
                 </div>
             </section>
         </article>
     </main>
     <footer>&copy; 2023 Melanie Archer (@archemel)</footer>
-    <script src="client.js"></script>
+    <script type="module" src="./modules/zipcode.js"></script>
+    <script type="module" src="./modules/places.js"></script>
+    <script type="module" src="./modules/client.js"></script>
 </body>
 
 </html>
